+++ 00_Base/src/index.ts
@@ -46,12 +46,11 @@ export {
   MessageState,
   RetryMessageError,
 } from './interfaces/messages';
-// export { AbstractModule, AsHandler, IModule } from './interfaces/modules';
-// export type { CircuitBreakerState, CircuitBreakerOptions } from './interfaces/modules';
 export {
   AbstractModule,
   AsHandler,
   IModule,
+  CircuitBreaker,
   type CircuitBreakerState,
   type CircuitBreakerOptions,
 } from './interfaces/modules';
@@ -257,4 +256,3 @@ export { AuthorizationSecurity } from './interfaces/api/AuthorizationSecurity';
 export { Ajv };
 export * from './interfaces/dto';
 export declare type Constructable<T> = new (...args: any[]) => T;
-export { CircuitBreaker } from './util/CircuitBreaker';
