+++ 00_Base/src/interfaces/modules/index.ts
@@ -7,4 +7,4 @@ export { AbstractModule } from './AbstractModule';
 export { AsHandler, AS_HANDLER_METADATA } from './AsHandler';
 export { IHandlerDefinition } from './HandlerDefinition';
 export { IModule } from './Module';
-export type { CircuitBreakerState, CircuitBreakerOptions } from './CircuitBreaker';
+export { CircuitBreaker, CircuitBreakerState, CircuitBreakerOptions } from './CircuitBreaker';
